conversation ok

+++ fp-gen-infrastructure/src/main/java/be/unamur/fpgen/repository/JpaInterlocutorRepositoryCRUD.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface JpaInterlocutorRepositoryCRUD extends JpaRepository<InterlocutorEntity, Long>{
+public interface JpaInterlocutorRepositoryCRUD extends JpaRepository<InterlocutorEntity, Integer>{
 
     List<InterlocutorEntity> findByInterlocutorTypeEnum(final InterlocutorTypeEnum type);
 }
